# docs/swagger.yaml
basePath: /api
definitions:
...
  controller.GetPrivateMessagesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Message'
        type: array
      size:
        type: integer
      total:
        type: integer
    type: object
  controller.PutRequesetRequestResponse:
    properties:
      status:
        description: 状态 `status` tinyint(1) DEFAULT '0' COMMENT '0-待确认, 1-已确认, 2-已拒绝,
          3-已删除',
        type: integer
    type: object
...
  controller.RequestResponse:
    properties:
      friendID:
        description: 好友的 用户ID (类型 uint)
        type: integer
      remark:
        description: 备注信息
        type: string
    type: object
  controller.ReseutPostGroupMember:
    properties:
      memberIDs:
        items:
          type: integer
        type: array
    type: object
...
  model.ChannelType:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-varnames:
    - ChannelTypeFriend
    - ChannelTypeGroup
    - ChannelTypePING
    - ChannelTypePONG
  model.Element:
    properties:
      content:
        description: 消息内容（文本内容或文件 HASH）
        type: string
      hash:
        description: 文件信息（仅在 Type 为图片/视频/文件时有效）
        type: string
      height:
        type: integer
      name:
        type: string
      size:
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/model.MessageType'
        description: 消息类型（0：文本，1：图片，2：视频，3：文件）
      url:
        type: string
      width:
        type: integer
    type: object
  model.Friend:
    properties:
      createdAt:
        type: string
      friendId:
        description: 好友的 用户ID (类型 uint)
        type: integer
      id:
        description: 唯一 ID
        type: integer
      remark:
        description: 发起好友申请时备注时的备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/model.RequestFriendStatus'
        description: '状态:  `status` tinyint(1) DEFAULT ''0'' COMMENT ''0-待确认, 1-已确认,
          2-已拒绝, 3-已删除'','
      updatedAt:
        type: string
      userId:
        description: 自身的 ID 和当前登录的 用户 ID 一直。 (类型 uint)
        type: integer
    type: object
  model.FriendResponse:
    properties:
      avatar:
        description: 用户头像
        type: string
      createdAt:
        type: string
      friendId:
        description: 好友的 用户ID (类型 uint)
        type: integer
      id:
        description: 唯一 ID
        type: integer
      isOnline:
        description: 在线状态
        type: boolean
      nickname:
        description: 用户中文名称
        type: string
      remark:
        description: 发起好友申请时备注时的备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/model.RequestFriendStatus'
        description: '状态:  `status` tinyint(1) DEFAULT ''0'' COMMENT ''0-待确认, 1-已确认,
          2-已拒绝, 3-已删除'','
      updatedAt:
        type: string
      userId:
        description: 自身的 ID 和当前登录的 用户 ID 一直。 (类型 uint)
        type: integer
      username:
        description: 用户账号
        type: string
    type: object
  model.GroupChat:
    properties:
      avatar:
        description: 群头像
        type: string
      createdAt:
        description: 创建时间
        type: string
      encrypted:
        description: 是否加密 0-否 1-是
        type: integer
      id:
        description: 群ID
        type: integer
      isMuted:
        description: 是否全员禁言 0-否 1-是
        type: integer
      name:
        description: 群名称
        type: string
      notice:
        description: 群公告
        type: string
      ownerId:
        description: 群主ID
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  model.GroupChatResponse:
    properties:
      avatar:
        description: 群头像
        type: string
      createdAt:
        description: 创建时间
        type: string
      encrypted:
        description: 是否加密 0-否 1-是
        type: integer
      id:
        description: 群ID
        type: integer
      isMuted:
        description: 是否全员禁言 0-否 1-是
        type: integer
      members:
        description: |-
          用户 ID；如果 ownerId == userId，则表示该用户是群主
          UserID UserID `gorm:"column:user_id;index;uniqueIndex:uk_group_user" json:"userId" swaggertype:"integer"`
          该群的所有成员信息；注意：该字段是通过关联查询获取的，而不是直接存储在数据库中的
        items:
          $ref: '#/definitions/model.GroupMemberResponse'
        type: array
      name:
        description: 群名称
        type: string
      notice:
        description: 群公告
        type: string
      ownerId:
        description: 群主ID
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  model.GroupMember:
    properties:
      createdAt:
        description: 创建时间
        type: string
      groupId:
        description: 群ID
        type: integer
      id:
        description: 成员ID
        type: integer
      isMuted:
        description: 该用户是否被禁言 0-否 1-是
        type: integer
      isNotifyDisabled:
        description: 该用户是否关闭该群通知 0-否 1-是
        type: integer
      joinedAt:
        description: 加入时间
        type: string
      role:
        description: 角色 0-成员 1-管理员 2-群主
        type: integer
      status:
        description: 该群员状态 0-正常 1-退群(自己退、被踢出)
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      userId:
        description: 用户ID
        type: integer
    type: object
  model.GroupMemberResponse:
    properties:
      avatar:
        description: 用户头像、创建时；为空字符串;注意只是保存 文件的hash 地址；而不是 URL 地址
        example: ''''''
        type: string
      createdAt:
        description: 创建时间
        type: string
      groupId:
        description: 群ID
        type: integer
      id:
        description: 成员ID
        type: integer
      isMuted:
        description: 该用户是否被禁言 0-否 1-是
        type: integer
      isNotifyDisabled:
        description: 该用户是否关闭该群通知 0-否 1-是
        type: integer
      joinedAt:
        description: 加入时间
        type: string
      nickname:
        description: 用户中文名称，默认为 ‘’，可后期通过修改用户信息设置
        type: string
      role:
        description: 角色 0-成员 1-管理员 2-群主
        type: integer
      status:
        description: 该群员状态 0-正常 1-退群(自己退、被踢出)
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      userId:
        description: 用户ID
        type: integer
      username:
        description: 用户账户；创建时使用；后期不可修改。
        example: whx
        type: string
    required:
    - avatar
    - username
    type: object
  model.Message:
    properties:
      createdAt:
        type: string
      elements:
        description: 消息元素列表（JSON 序列化存储）
        items:
          $ref: '#/definitions/model.Element'
        type: array
      id:
        description: 消息 数据库索引
        type: integer
      receiverId:
        description: 消息接收者 ID, 如果是好友消息则为好友的 UserID，如果是群组消息则为群组的 GroupID
        type: integer
      senderId:
        description: 消息发送者 ID
        type: integer
      status:
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/model.ChannelType'
        description: 消息类型 (0：好友消息，1：群组消息，2：PING，3：PONG)
    type: object
  model.MessageType:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - MsgTypeText
    - MsgTypeImage
    - MsgTypeVideo
    - MsgTypeFile
...
  model.RequestFriendStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-comments:
      FriendStatusAccepted: 已确认
      FriendStatusDeleted: 已删除
      FriendStatusPending: 待确认
      FriendStatusRejected: 已拒绝
    x-enum-descriptions:
    - 待确认
    - 已确认
    - 已拒绝
    - 已删除
    x-enum-varnames:
    - FriendStatusPending
    - FriendStatusAccepted
    - FriendStatusRejected
    - FriendStatusDeleted
...
  utils.KResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    一个 IM 服务器
    支持用户注册、登录、好友管理、群组聊天
    基于 Go + Gin + GORM + Redis
    目前访问服务器任何接口: localhost:8080/api 前端可用先使用代理设置
    使用文件 hash 获取文件地址可用先固定: localhost:8080/api/file/:hash 前端可用看作做;方便后期迁移
  title: go-net
  version: "1.0"
paths:
...
  /friends/{id}:
    put:
      description: 如果将好友移除；同时也会移除对方的好友关系
      parameters:
      - description: 修改好友申请状态;好友列表（表的 ID）
        in: path
        name: id
        required: true
        type: integer
      - description: 修改好友申请状态
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.PutRequesetRequestResponse'
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/controller.PutRequesetRequestResponse'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 修改好友申请状态,可用于 同意(1),拒绝(2)、删除(3)
      tags:
      - 好友模块
  /friends/friends:
    get:
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.FriendResponse'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 获取好友列表
      tags:
      - 好友模块
  /friends/request:
    post:
      description: |-
        好友表没有关系则新增一条记录 status = 0
        好友表中已经存在；且 status != 1 或 != 0时；则重新更新 status = 0 和 remark(如果有传递)
        好友表中已经存在；且 status == 1 或 = 0时；不会有任何变化；返回成功
      parameters:
      - description: 发起好友申请
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.RequestResponse'
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Friend'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 发起好友申请
      tags:
      - 好友模块
  /friends/requests:
    get:
      description: user_id = 当前登录的代表主动发起申请(等待别人反馈)；否则；是其他用户(friend_id)向当前用户发起生气 [同意，拒绝]
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.FriendResponse'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 获取申请列表
      tags:
      - 好友模块
  /groups:
    get:
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.GroupChatResponse'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 获取群列表
      tags:
      - 群
  /groups/:id/members:
    post:
      parameters:
      - description: 群ID
        in: path
        name: id
        required: true
        type: integer
      - description: 成员ID列表(用户 ID)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ReseutPostGroupMember'
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.GroupChat'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 给该群添加群成员
      tags:
      - 群
  /groups/:id/putMember:
    put:
      description: 注意： 所属群主无法修改 Role、Status 状态
      parameters:
      - description: 群ID
        in: path
        name: id
        required: true
        type: integer
      - description: 需要修改的数据；注意：UserID 和 ID 字段无效，系统会自动获取当前用户的 ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.GroupMember'
      responses:
        "200":
          description: 成功-修改成功后的数据
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.GroupMember'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 修改当前用户在该群的信息(如：是否静音、是否加密、是否退群等)
      tags:
      - 群
  /groups/{id}:
    get:
      parameters:
      - description: 群ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.GroupChatResponse'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 获取群详情
      tags:
      - 群
    put:
      parameters:
      - description: 群ID
        in: path
        name: id
        required: true
        type: integer
      - description: 群信息-传递 ID 字段无效
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.GroupChat'
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.GroupChat'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 修改群信息
      tags:
      - 群
  /messages/friend/:friendID:
    get:
      parameters:
      - default: 20
        description: 限制条数
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/controller.GetPrivateMessagesResponse'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 获取与好友的历史记录
      tags:
      - 消息
  /messages/group/:groupID:
    get:
      parameters:
      - default: 20
        description: 限制条数
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/controller.GetPrivateMessagesResponse'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 获取群组消息
      tags:
      - 消息
...
swagger: "2.0"